Make sure Item throws an exception and doesn't return null. Also, catch EOFExceptions in NetIOController.read

+++ SWEN 222 Game/src/game/exceptions/ItemIDNotFoundException.java
@@ -3,7 +3,7 @@ package game.exceptions;
 public class ItemIDNotFoundException extends GameException {
 	private static final long serialVersionUID = 1L;
 
-	public ItemIDNotFoundException(byte id) {
+	public ItemIDNotFoundException(int id) {
 		super(false, "item ID not found: " + id);
 	}
 }
